+++ attributes/default.rb
@@ -1,0 +1,3 @@
+default['spigot']['current_version'] = 'latest'
+default['minecraft']['version'] = 'latest'
+default['minecraft']['jar_source'] = ''