+++ recipes/default.rb
@@ -1,0 +1,1 @@
+# This is just an example recipe and should not be used. Please use the custom resources provided by this cookbook.