+++ resources/minecraft_service.rb
@@ -1,10 +1,11 @@
 resource_name :minecraft_service
 provides :minecraft_service
 
+property :additional_cmd, String, default: ''
 property :commands, [Array, String], default: []
 property :command_delay, [Integer, String], default: 2
-property :delay, Integer, default: 10
 property :group, String, default: 'chefminecraft'
+property :is_bungeecord, [TrueClass, FalseClass], default: false
 property :jar_name, String, default: 'minecraft_server.jar'
 property :owner, String, default: 'chefminecraft'
 property :path, String, default: '/opt/minecraft_servers'
@@ -13,17 +14,45 @@ property :service_name, String, name_property: 'default'
 default_action :nothing
 
 action_class do
-  # Helper methods here
+  include Minecraft_Server::Utils
+
+  def is_done_loading(file)
+    if ::File.exists?(file)
+      filelines = ::IO.readlines(file)
+      filelines.each do |line|
+        found = line.match(/Done \([0-9]*\.[0-9]*s\)\! For help, type "help" or "\?"/)
+        if found.nil?
+          found = line.match(/Listening on (\/[0-9.a-zA-Z]+):([0-9]+)/)
+          unless found.nil?
+            return true
+          end
+        else
+          return true
+        end
+      end
+      false
+    else
+      false
+    end
+  end
 end
 
 action :start do
   if ::File.exist?("/etc/systemd/system/minecraft_#{new_resource.service_name}.service") || ::File.exist?("/etc/init.d/minecraft_#{new_resource.service_name}")
-    service "minecraft_#{new_resource.service_name}.service" do
+    service "minecraft_#{new_resource.service_name}" do
       action :start
     end
     ruby_block 'start_wait' do
       block do
-        sleep(new_resource.delay)
+        count = 0
+        puts ''
+        while is_running(new_resource.service_name) && !is_done_loading("#{new_resource.path}/#{new_resource.service_name}/logs/latest.log") do
+          if count % 5 == 0
+            puts 'Waiting for the server to finish starting...'
+          end
+          count += 1
+          sleep(1)
+        end
       end
     end
     ruby_block 'execute commands' do
@@ -65,12 +94,22 @@ action :stop do
       end
     end
 
-    service "minecraft_#{new_resource.service_name}.service" do
+    service "minecraft_#{new_resource.service_name}" do
+      service_name "minecraft_#{new_resource.service_name}"
       action :stop
     end
+
     ruby_block 'stop_wait' do
       block do
-        sleep(new_resource.delay)
+        count = 0
+        puts ''
+        while is_running(new_resource.service_name) do
+          if count % 5 == 0
+            puts 'Waiting for the server to shut down...'
+          end
+          count += 1
+          sleep(1)
+        end
       end
     end
   else
@@ -89,12 +128,30 @@ action :restart do
     end
     ruby_block 'restart_wait' do
       block do
-        sleep(new_resource.delay)
+        count = 0
+        puts ''
+        while is_running(new_resource.service_name) do
+          if count % 5 == 0
+            puts 'Waiting for the server to shut down...'
+          end
+          count += 1
+          sleep(1)
+        end
       end
     end
     service "minecraft_#{new_resource.service_name}_start" do
       service_name "minecraft_#{new_resource.service_name}.service"
       action :start
+    end
+
+    count = 0
+    puts ''
+    while is_running(new_resource.service_name) && !is_done_loading("#{new_resource.path}/#{new_resource.service_name}/logs/latest.log") do
+      if count % 5 == 0
+        puts 'Waiting for the server to finish starting...'
+      end
+      count += 1
+      sleep(1)
     end
   else
     log 'no service' do
@@ -106,7 +163,7 @@ end
 
 action :enable do
   if ::File.exist?("/etc/systemd/system/minecraft_#{new_resource.service_name}.service") || ::File.exist?("/etc/init.d/minecraft_#{new_resource.service_name}")
-    service "minecraft_#{new_resource.service_name}.service" do
+    service "minecraft_#{new_resource.service_name}" do
       action :enable
     end
   else
@@ -119,7 +176,7 @@ end
 
 action :disable do
   if ::File.exist?("/etc/systemd/system/minecraft_#{new_resource.service_name}.service") || ::File.exist?("/etc/init.d/minecraft_#{new_resource.service_name}")
-    service "minecraft_#{new_resource.service_name}.service" do
+    service "minecraft_#{new_resource.service_name}" do
       action :disable
     end
   else
@@ -139,11 +196,18 @@ action :create do
         variables({
                       :user => new_resource.owner,
                       :name => new_resource.service_name,
-                      :jar_name => "#{new_resource.jar_name}}.jar",
-                      :directory => "#{new_resource.path}/#{new_resource.service_name}"
+                      :jar_name => "#{new_resource.jar_name}.jar",
+                      :directory => "#{new_resource.path}/#{new_resource.service_name}",
+                      :stop_command => new_resource.is_bungeecord ? 'end' : 'stop',
+                      :add_cmd => new_resource.additional_cmd.eql?('') ? '' : " #{new_resource.additional_cmd}",
+                      :init_function => '/etc/init.d/functions'
                   })
         mode '0755'
         not_if { ::File.exist?("/etc/init.d/minecraft_#{new_resource.service_name}") }
+      end
+
+      execute 'add script to chkconfig' do
+        command "chkconfig --add minecraft_#{new_resource.service_name}"
       end
     elsif node['platform_version'].slice(0, 1).eql? '7'
       template "/etc/systemd/system/minecraft_#{new_resource.service_name}.service" do
@@ -154,7 +218,9 @@ action :create do
                       :group => new_resource.group,
                       :name => new_resource.service_name,
                       :jar_name => "#{new_resource.jar_name}.jar",
-                      :directory => "#{new_resource.path}/#{new_resource.service_name}"
+                      :directory => "#{new_resource.path}/#{new_resource.service_name}",
+                      :stop_command => new_resource.is_bungeecord ? 'end' : 'stop',
+                      :add_cmd => new_resource.additional_cmd.eql?('') ? '' : " #{new_resource.additional_cmd}"
                   })
         mode '0755'
         not_if { ::File.exist?("/etc/systemd/system/minecraft_#{new_resource.service_name}.service") }
@@ -181,7 +247,10 @@ action :create do
                   :user => new_resource.owner,
                   :name => new_resource.service_name,
                   :jar_name => "#{new_resource.jar_name}.jar",
-                  :directory => "#{new_resource.path}/#{new_resource.service_name}"
+                  :directory => "#{new_resource.path}/#{new_resource.service_name}",
+                  :stop_command => new_resource.is_bungeecord ? 'end' : 'stop',
+                  :add_cmd => new_resource.additional_cmd.eql?('') ? '' : " #{new_resource.additional_cmd}",
+                  :init_function => '/lib/lsb/init-functions'
               }
             }
         )
@@ -197,7 +266,9 @@ action :create do
                       :group => new_resource.group,
                       :name => new_resource.service_name,
                       :jar_name => "#{new_resource.jar_name}.jar",
-                      :directory => "#{new_resource.path}/#{new_resource.service_name}"
+                      :directory => "#{new_resource.path}/#{new_resource.service_name}",
+                      :stop_command => new_resource.is_bungeecord ? 'end' : 'stop',
+                      :add_cmd => new_resource.additional_cmd.eql?('') ? '' : " #{new_resource.additional_cmd}"
                   })
         mode '0755'
         not_if { ::File.exist?("/etc/systemd/system/minecraft_#{new_resource.service_name}.service") }
@@ -225,8 +296,11 @@ action :update do
         variables({
                       :user => new_resource.owner,
                       :name => new_resource.service_name,
-                      :jar_name => "#{new_resource.jar_name}}.jar",
-                      :directory => "#{new_resource.path}/#{new_resource.service_name}"
+                      :jar_name => "#{new_resource.jar_name}.jar",
+                      :directory => "#{new_resource.path}/#{new_resource.service_name}",
+                      :stop_command => new_resource.is_bungeecord ? 'end' : 'stop',
+                      :add_cmd => new_resource.additional_cmd.eql?('') ? '' : " #{new_resource.additional_cmd}",
+                      :init_function => '/etc/init.d/functions'
                   })
         mode '0755'
       end
@@ -239,7 +313,9 @@ action :update do
                       :group => new_resource.group,
                       :name => new_resource.service_name,
                       :jar_name => "#{new_resource.jar_name}.jar",
-                      :directory => "#{new_resource.path}/#{new_resource.service_name}"
+                      :directory => "#{new_resource.path}/#{new_resource.service_name}",
+                      :stop_command => new_resource.is_bungeecord ? 'end' : 'stop',
+                      :add_cmd => new_resource.additional_cmd.eql?('') ? '' : " #{new_resource.additional_cmd}"
                   })
         mode '0755'
       end
@@ -264,7 +340,10 @@ action :update do
                 :user => new_resource.owner,
                 :name => new_resource.service_name,
                 :jar_name => "#{new_resource.jar_name}.jar",
-                :directory => "#{new_resource.path}/#{new_resource.service_name}"
+                :directory => "#{new_resource.path}/#{new_resource.service_name}",
+                :stop_command => new_resource.is_bungeecord ? 'end' : 'stop',
+                :add_cmd => new_resource.additional_cmd.eql?('') ? '' : " #{new_resource.additional_cmd}",
+                :init_function => '/lib/lsb/init-functions'
             }
         )
         mode '0755'
@@ -278,7 +357,9 @@ action :update do
                       :group => new_resource.group,
                       :name => new_resource.service_name,
                       :jar_name => "#{new_resource.jar_name}.jar",
-                      :directory => "#{new_resource.path}/#{new_resource.service_name}"
+                      :directory => "#{new_resource.path}/#{new_resource.service_name}",
+                      :stop_command => new_resource.is_bungeecord ? 'end' : 'stop',
+                      :add_cmd => new_resource.additional_cmd.eql?('') ? '' : " #{new_resource.additional_cmd}"
                   })
         mode '0755'
       end