+++ resources/spigot_server.rb
@@ -18,6 +18,8 @@ property :world, String, default: ''
 default_action :create
 
 action_class do
+  include Minecraft_Server::Utils
+
   def get_spigot_version
     unless ::File.exist?("#{new_resource.build_tools_dir}/working_version.txt")
       Chef::Application.fatal!("Failed to find the working_version.txt file in #{build_tools_dir}!", 1)
@@ -40,7 +42,7 @@ action_class do
     folders.each do |folderName|
       temp = find_world_file(folderName)
       unless temp.nil?
-        temp
+        return temp
       end
     end
     nil
@@ -168,16 +170,18 @@ chown -R #{new_resource.owner}:#{new_resource.group} world/
 end
 
 action :update do
+  running = is_running(new_resource.name)
   minecraft_service "#{new_resource.name}_stop" do
     service_name new_resource.name
     action :stop
+    only_if { running }
   end
-  if new_resource.jar_source.eql? ''
-    build_tools 'create jar' do
-      version new_resource.version
-      update_jar new_resource.update_jar
-      action :build
-    end
+
+  build_tools 'create jar' do
+    version new_resource.version
+    update_jar new_resource.update_jar
+    action :build
+    only_if { new_resource.jar_source.eql? '' && new_resource.update_jar }
   end
 
   minecraft_service new_resource.name do
@@ -190,6 +194,7 @@ action :update do
     end
     path new_resource.path
     action :update
+    only_if { new_resource.update_jar }
   end
 
   ruby_block 'remove old jar' do
@@ -201,29 +206,32 @@ action :update do
           break
         end
       end
+      puts jar
       unless jar.eql? ''
-        if jar.eql? "spigot-#{node['spigot']['current_version']}.jar"
+        if jar.eql?("spigot-#{node['spigot']['current_version']}.jar")
           ::FileUtils.rm("#{new_resource.path}/#{new_resource.name}/#{jar}")
-        elsif jar.eql? 'spigot.jar'
+        elsif jar.eql?('spigot.jar')
           ::FileUtils.rm("#{new_resource.path}/#{new_resource.name}/#{jar}")
         end
       end
     end
+    only_if { new_resource.update_jar }
   end
 
-  if new_resource.jar_source.eql? ''
-    ruby_block 'copy jar' do
-      block do
-        ::FileUtils.cp("#{new_resource.build_tools_dir}/spigot-#{node['spigot']['current_version']}.jar", "#{new_resource.path}/#{new_resource.name}")
-      end
+
+  ruby_block 'copy jar' do
+    block do
+      ::FileUtils.cp("#{new_resource.build_tools_dir}/spigot-#{node['spigot']['current_version']}.jar", "#{new_resource.path}/#{new_resource.name}")
     end
-  else
-    remote_file "#{new_resource.path}/#{new_resource.name}/spigot.jar" do
-      source new_resource.jar_source
-      owner new_resource.owner
-      group new_resource.group
-      action :create
-    end
+    only_if { new_resource.jar_source.eql? '' && new_resource.update_jar }
+  end
+
+  remote_file "#{new_resource.path}/#{new_resource.name}/spigot.jar" do
+    source new_resource.jar_source
+    owner new_resource.owner
+    group new_resource.group
+    action :create
+    only_if { new_resource.update_jar && !new_resource.jar_source.eql?('') }
   end
 
   if new_resource.world.eql? ''
@@ -295,6 +303,7 @@ action :update do
   minecraft_service "#{new_resource.name}_start" do
     service_name new_resource.name
     action :start
+    only_if { running }
   end
 end