+++ test/integration/default/serverspec/default_spec.rb
@@ -1,9 +1,3 @@
 require 'spec_helper'
 
-describe 'minecraft-server::default' do
-  # Serverspec examples can be found at
-  # http://serverspec.org/resource_types.html
-  it 'does something' do
-    skip 'Replace this with meaningful tests'
-  end
-end
+# Eventually do something?